add outfit suggestion in navbar
